refactor: move _check_dependencies to taipy.common
