+++ taipy/core/_entity/_reload.py
@@ -12,8 +12,8 @@
 import functools
 from typing import Dict, Type
 
+from ...common._check_dependencies import EnterpriseEditionUtils
 from .._manager._manager import _Manager
-from ..common._check_dependencies import EnterpriseEditionUtils
 from ..common._utils import _load_fct
 from ..notification import EventOperation, Notifier, _make_event
