+++ taipy/core/_orchestrator/_orchestrator_factory.py
@@ -8,12 +8,13 @@
 # Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
 # an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
 # specific language governing permissions and limitations under the License.
+
 import typing
 from typing import Optional, Type
 
 from taipy.common.config import Config
 
-from ..common._check_dependencies import EnterpriseEditionUtils
+from ...common._check_dependencies import EnterpriseEditionUtils
 from ..common._utils import _load_fct
 from ..exceptions.exceptions import ModeNotAvailable, OrchestratorNotBuilt
 from ._abstract_orchestrator import _AbstractOrchestrator
