+++ taipy/core/_version/_cli/_version_cli_factory.py
@@ -15,7 +15,7 @@ from typing import Type
 
 from taipy.common._cli._base_cli._abstract_cli import _AbstractCLI
 
-from ...common._check_dependencies import EnterpriseEditionUtils
+from ....common._check_dependencies import EnterpriseEditionUtils
 from ._version_cli import _VersionCLI
