+++ taipy/core/sequence/_sequence_manager_factory.py
@@ -12,8 +12,8 @@
 from functools import lru_cache
 from typing import Type
 
+from ...common._check_dependencies import EnterpriseEditionUtils
 from .._manager._manager_factory import _ManagerFactory
-from ..common._check_dependencies import EnterpriseEditionUtils
 from ..common._utils import _load_fct
 from ._sequence_manager import _SequenceManager
